IN-314 refactor: refactoring
