--- src/pages/profile/[id]/publications/[postId]/view/index.ts
@@ -1,3 +1,0 @@
-import PostImageView from '../../../../../../features/gradient/PostImageView'
-
-export default PostImageView
