+++ src/pages/profile/[id]/publications/[postId]/view/index.tsx
@@ -1,0 +1,3 @@
+import { PostImageView } from '@/features/post-image-view/ui/PostImageView'
+
+export default PostImageView
