+++ src/widgets/post-details/ui/PostDetails.tsx
@@ -2,9 +2,8 @@ import React from 'react'
 
 import { commentsApi } from '@/entities/comments'
 import { Post } from '@/entities/posts'
-import { Button } from '@byte-creators/ui-kit'
-import { EyeOutline } from '@byte-creators/ui-kit/icons'
-import { cn, useMediaQuery } from '@byte-creators/utils'
+import { PostDetailsSlide } from '@/widgets/post-details/ui/PostDetailsSlide'
+import { useMediaQuery } from '@byte-creators/utils'
 import { useRouter } from 'next/router'
 
 import { PostDesktop } from './desktop/PostDesktop'
@@ -24,24 +23,8 @@ export const PostDetails = ({ post }: Props) => {
       query: { image: proxyUrl },
     })
   }
-  //TODO: remove any
   const slides = post.images.map((image: any, i) => (
-    <div className={'w-full h-full relative'} key={i}>
-      <img alt={'postImg'} className={'h-full object-cover object-center sc'} src={image.url} />
-      <Button
-        className={cn(
-          'text-2xl absolute inset-0 m-auto opacity-0',
-          'transition-opacity duration-300 hover:opacity-100',
-          'px-36 focus:outline-none active:outline-none !outline-none'
-        )}
-        onClick={() => handleNavigateToImage(image.url)}
-        variant={'text'}
-      >
-        <div className={'bg-dark-100 rounded-full inline-block p-2'}>
-          <EyeOutline className={'text-light-100 transform scale-150'} />
-        </div>
-      </Button>
-    </div>
+    <PostDetailsSlide handleNavigateToImage={handleNavigateToImage} image={image} key={i} />
   ))
   const isLargeScreen = useMediaQuery('(min-width: 768px)')
